PlayerIDs, and working bid validation.
 * Use PlayerID everywhere
 * new AllBids class to encapsulate this better
 * Update Player interface a bit
 * Give up and use Integers for bids
 * Add new boring but non-random bidding strategy
 * Add tests

+++ src/main/java/net/declension/games/cards/tricks/BidAndTaken.java
@@ -1,10 +1,10 @@
 package net.declension.games.cards.tricks;
 
 public class BidAndTaken {
-    private final short bid;
-    private final short score;
+    private final int bid;
+    private final int score;
 
-    public BidAndTaken(short bid, short score) {
+    public BidAndTaken(int bid, int score) {
         this.bid = bid;
         this.score = score;
     }
